requirements.txt
Fabric==1.8.2
Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.4
Jinja2==2.7.2
MarkupSafe==0.18
Pillow==2.7.0
SQLAlchemy==0.9.3
WTForms==1.0.5
Werkzeug==0.9.4
argparse==1.2.1
coverage==3.7.1
ecdsa==0.10
itsdangerous==0.23
mechanize==0.2.5
nltk==3.0.0
paramiko==1.12.2
pycrypto==2.6.1
pytesseract==0.1.5
requests==2.5.1
wsgiref==0.1.2